Move remaining classes out of internal
