+++ platform-sdk/consensus-concurrent/src/main/java/module-info.java
@@ -6,7 +6,6 @@ module org.hiero.consensus.concurrent {
     exports org.hiero.consensus.concurrent.config;
     exports org.hiero.consensus.concurrent.framework;
     exports org.hiero.consensus.concurrent.framework.config;
-    exports org.hiero.consensus.concurrent.framework.internal;
     exports org.hiero.consensus.concurrent.framework.queue;
     exports org.hiero.consensus.concurrent.manager;
     exports org.hiero.consensus.concurrent.pool;
